watched video done

+++ app/src/main/java/tube/chikichiki/sako/model/WatchedVideo.kt
@@ -4,5 +4,6 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 import java.util.UUID
 
+//This is used for progress bar under video thumbnails to show how much the user previously watched of the video
 @Entity
-data class WatchedVideo(@PrimaryKey val uuid: UUID, var watchedVideoTimeInMil:Float)
+data class WatchedVideo(@PrimaryKey val uuid: UUID, var watchedVideoTimeInMil:Long)
